Reuse old ThreadData->funcAddr (now funcAddr_or_gcEnv) to alloc GC obj

+++ runtime/gc.c
@@ -1,8 +1,9 @@
 
+#include <stdint.h>
 #include <stdlib.h>
 #include "gc.h"
 
-void* __GC_malloc(size_t alloc_size, GC_Env* gc_env)
+void* __GC_malloc(uint64_t alloc_size, GC_Env* gc_env)
 {
     void* alloc = malloc(alloc_size);
     add_alloc(alloc, gc_env);
@@ -13,7 +14,7 @@ void add_alloc(void* alloc, GC_Env* gc_env)
 {
     if (gc_env->allocs == NULL)
     {
-        size_t start_size = ALLOCS_START_SIZE;
+        uint64_t start_size = ALLOCS_START_SIZE;
         void** allocs = malloc(start_size);
         if (allocs == NULL)
         {
@@ -25,7 +26,7 @@ void add_alloc(void* alloc, GC_Env* gc_env)
     }
     else if (gc_env->allocs_len >= gc_env->allocs_end)
     {
-        size_t new_size = gc_env->allocs_end * 2;
+        uint64_t new_size = gc_env->allocs_end * 2;
         void** new_allocs = realloc(gc_env->allocs, new_size);
         if (new_allocs == NULL)
         {
@@ -40,7 +41,7 @@ void add_alloc(void* alloc, GC_Env* gc_env)
 
 void free_all_allocs(GC_Env* gc_env)
 {
-    size_t i;
+    uint64_t i;
     for (i = 0; i < gc_env->allocs_len; i++)
     {
         free(gc_env->allocs[i]);
