+++ runtime/gc.h
@@ -2,12 +2,14 @@
 #ifndef GC_H
 #define GC_H
 
+#include <stdint.h>
+
 #define ALLOCS_START_SIZE 64
 
 typedef struct {
     void** allocs;
-    size_t allocs_len;
-    size_t allocs_end;
+    uint64_t allocs_len;
+    uint64_t allocs_end;
 } GC_Env;
 
 #endif
